chore(ci): husky hooks + fix clippy + drop stale dmc-core integration tests
CI was failing on `cargo clippy --workspace --all-targets -- -D warnings`
and on missing fixture paths in dmc-core integration tests. Fix the
clippy errors, drop the broken integration tests (they used the
pre-refactor `dmc::compile(src)` / `dmc::run` / `dmc::EngineConfig` API
that no longer exists; recreate against the current `Compiler::compile`
+ `Engine::run` shape later), and stand up local pre-commit / pre-push
hooks that mirror the CI checks so the next breakage is caught locally.

Husky:
- New `.husky/pre-commit`: `cargo fmt --all` (apply, never reject).
- New `.husky/pre-push`: `cargo fmt --all -- --check`,
  `cargo clippy --workspace --all-targets -- -D warnings`, then
  `cargo test --workspace`. Push fails fast if anything is dirty.
- `package.json`: add `husky@^9` dev dep + scripts (`prepare`, `fix`,
  `check`, `test`). One-time setup: `npm run prepare` after clone.

Clippy fixes:
- dmc-transform/src/builtin/copy_linked_files.rs: collapse nested `if`
  inside `match Node::Link` into the arm guard.
- dmc-transform/src/pipeline.rs: drop redundant `&mut` on `engine`.
- dmc-parser/tests/structure.rs: remove unused `pretty_assertions::assert_eq` import.
- dmc-schema/tests/basic.rs: drop unneeded `mut` on Ctx (parse takes &self, &Ctx).
- dmc-transform/tests/{component_source,disable_gfm}.rs: drop unused `Transformer` import.
- dmc-core/src/cli/dev.rs: `&PathBuf` -> `&Path` in `rebuild`.
- dmc-core/src/engine/accumlator.rs: add `Default` impl.
- dmc-core/src/engine/collection.rs: replace `is_some` + `unwrap` with `if let Some`.
- dmc-core/examples/bench.rs: tighten doc-list indent, drop redundant `as i32` casts.

Stale tests (pre-existing breakage from earlier API refactor):
- Drop `dmc-core/tests/{cli,compile_basic,engine,mdx_options,parity,robustness,snapshots}.rs`
  + the `snapshots/` snap dir. They referenced `dmc::compile(src)` /
  `dmc::run` / `dmc::EngineConfig` etc, which moved to
  `dmc::engine::compile::Compiler::compile` and friends. Rewrite later.
- `dmc-core/benches/compile.rs`: replace missing-fixture
  `include_str!("../../tests/fixtures/velite-parity/skills.mdx")` with
  an inline FIXTURE constant; criterion benches now compile + run
  without the deleted tests/fixtures/ tree.

+++ dmc-core/examples/bench.rs
@@ -5,8 +5,8 @@
 //!   2. Full build, native (cold)        - wall time vs N files.
 //!   3. Full build, sidecar light        - + remark-gfm.
 //!   4. Full build, sidecar heavy        - + remark-gfm + rehype-pretty-code (shiki).
-//!   5. Full build, sidecar kitchen-sink - gfm + math + katex + emoji +
-//!                                         pretty-code + slug + autolink-headings.
+//!   5. Full build, sidecar kitchen-sink - gfm + math + katex + emoji
+//!      + pretty-code + slug + autolink-headings.
 //!   6. Full build, velite light         - reference (gfm only).
 //!   7. Full build, velite kitchen-sink  - reference (same plugin chain).
 //!
@@ -1045,7 +1045,7 @@ fn plot_table(out: &Path, report: &Report) -> Result<(), Box<dyn std::error::Err
   for (ci, h) in headers.iter().enumerate() {
     let cw = col_w[ci] as i32;
     root.draw(&Rectangle::new([(x, y0), (x + cw, y0 + row_h)], border.stroke_width(1)))?;
-    root.draw_text(h, &header_font, (x + cell_pad as i32, y0 + 9))?;
+    root.draw_text(h, &header_font, (x + cell_pad, y0 + 9))?;
     x += cw;
   }
 
@@ -1070,7 +1070,7 @@ fn plot_table(out: &Path, report: &Report) -> Result<(), Box<dyn std::error::Err
       } else {
         cell_font.clone()
       };
-      root.draw_text(cell, &style, (x + cell_pad as i32, yy + 9))?;
+      root.draw_text(cell, &style, (x + cell_pad, yy + 9))?;
       x += cw;
     }
   }
